requirements.txt
Package              Version
-------------------- -------
asgiref              3.4.1
Django               3.2.7
django-crispy-forms  1.12.0
django-shopping-cart 0.1
Pillow               8.3.2
pip                  21.2.4
pytz                 2021.1
setuptools           47.1.0
sqlparse             0.4.2
